Updated the schema with the suggested changes
